+++ CHANGES.md
@@ -8,6 +8,7 @@ History
 * Change `CDPSession.send` to a normal function which returns awaitable value
 * Add `Page.isClosed` method
 * Add `ElementHandle.querySelectorAllEval` and `ElementHandle.JJeval`
+* Add `Target.opener`
 
 ## Version 0.0.22 (2018-09-06)